Repository/code/es-dumper/requirements.txt
elasticsearch
ruamel.yaml